backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9

# LangGraph & LangChain
langgraph==0.0.32
langchain==0.1.16
langchain-google-genai==1.0.1
langchain-core==0.1.42

# Google Gemini
google-generativeai==0.4.1

# 데이터 처리
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.0
python-multipart==0.0.9

# 유사도 검색 (RAG)
scikit-learn==1.4.1.post1
numpy==1.26.4

# HTTP 클라이언트
requests==2.31.0
httpx==0.27.0

# 유틸리티
tenacity==8.2.3  # 재시도 로직
python-dateutil==2.8.2
